Phase 3: Refactor core configs
- Move configs/openai_api_models.py → core/configs/openai_api_models.py
- Move configs/server_models.py → core/configs/llm_config.py
- Move configs/constants.py → constants.py (serve level)
- Update all imports across codebase including public API
- All tests passing (43 tests validated)
