+++ python/ray/llm/_internal/common/callbacks/base.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Any, Callable, Dict, Optional, Tuple, Type, Un
 
 if TYPE_CHECKING:
     from ray.llm._internal.common.utils.download_utils import NodeModelDownloadable
-    from ray.llm._internal.serve.configs.server_models import LLMConfig
+    from ray.llm._internal.serve.core.configs.llm_config import LLMConfig
 
 logger = logging.getLogger(__name__)
