+++ python/ray/llm/_internal/serve/deployments/utils/server_utils.py
@@ -8,7 +8,7 @@ from httpx import HTTPStatusError as HTTPXHTTPStatusError
 from pydantic import ValidationError as PydanticValidationError
 
 from ray import serve
-from ray.llm._internal.serve.configs.openai_api_models import (
+from ray.llm._internal.serve.core.configs.openai_api_models import (
     ErrorInfo,
     ErrorResponse,
     OpenAIHTTPException,
