+++ python/ray/llm/tests/serve/conftest.py
@@ -11,7 +11,7 @@ import yaml
 
 import ray
 from ray import serve
-from ray.llm._internal.serve.configs.openai_api_models import (
+from ray.llm._internal.serve.core.configs.openai_api_models import (
     ChatCompletionRequest,
     CompletionRequest,
     EmbeddingCompletionRequest,
