+++ python/ray/llm/tests/serve/cpu/configs/test_models.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import pydantic
 import pytest
 
-from ray.llm._internal.serve.configs.server_models import (
+from ray.llm._internal.serve.core.configs.llm_config import (
     LLMConfig,
     LoraConfig,
     ModelLoadingConfig,
