+++ python/ray/llm/tests/serve/cpu/deployments/data_parallel/test_dp_server.py
@@ -3,7 +3,7 @@ from copy import deepcopy
 
 import pytest
 
-from ray.llm._internal.serve.configs.server_models import LLMConfig
+from ray.llm._internal.serve.core.configs.llm_config import LLMConfig
 from ray.llm._internal.serve.deployments.data_parallel.dp_server import DPServer
