+++ python/ray/llm/tests/serve/cpu/deployments/llm/test_builder_llm_server.py
@@ -3,7 +3,7 @@ import sys
 import pytest
 
 from ray import serve
-from ray.llm._internal.serve.configs.server_models import (
+from ray.llm._internal.serve.core.configs.llm_config import (
     LLMConfig,
 )
 from ray.llm._internal.serve.deployments.llm.builder_llm_server import (
