+++ python/ray/llm/tests/serve/cpu/deployments/llm/test_llm_server.py
@@ -8,7 +8,7 @@ import numpy as np
 import pytest
 
 from ray import serve
-from ray.llm._internal.serve.configs.server_models import (
+from ray.llm._internal.serve.core.configs.llm_config import (
     LLMConfig,
     LoraConfig,
     ModelLoadingConfig,
