+++ python/ray/llm/tests/serve/cpu/deployments/routers/test_router.py
@@ -6,7 +6,7 @@ import openai
 import pytest
 
 from ray import serve
-from ray.llm._internal.serve.configs.server_models import (
+from ray.llm._internal.serve.core.configs.llm_config import (
     LLMConfig,
     ModelLoadingConfig,
 )
