+++ python/ray/llm/tests/serve/mocks/mock_vllm_engine.py
@@ -5,7 +5,7 @@ from random import randint
 from typing import AsyncGenerator, Dict, Union
 
 from ray.llm._internal.common.utils.cloud_utils import LoraMirrorConfig
-from ray.llm._internal.serve.configs.openai_api_models import (
+from ray.llm._internal.serve.core.configs.openai_api_models import (
     ChatCompletionRequest,
     ChatCompletionResponse,
     CompletionRequest,
@@ -16,7 +16,7 @@ from ray.llm._internal.serve.configs.openai_api_models import (
     ScoreRequest,
     ScoreResponse,
 )
-from ray.llm._internal.serve.configs.server_models import (
+from ray.llm._internal.serve.core.configs.llm_config import (
     DiskMultiplexConfig,
     LLMConfig,
 )
