+++ python/requirements/llm/llm-test-requirements.txt
@@ -1,4 +1,5 @@
 pytest
+pytest-asyncio
 aiohttp
 pillow
 httpx>=0.27.2
